Add files via upload
Here are my "Remove Trap" trainers using the Circuit Trap Trainer box (either one you own, or the one at the community center. Just be aware that Ive confirmed that when others are using the same circuit trap trainer box, it can imact your use. So the best bet is to buy a trainer box yourself and train at home. 

After buying the skill to 40, I trained to 80 with the 3x3 trainer in 3 days while working my job (not AFK training, just multiple PC's at the same desk) and then trained to 110 skill in 2 days using the 4x4 trainer script. 

Known Issues: Though the scripts are smart enough to account for server lag, very occasionally one of the scripts will hang due to a server save. When that happens you must restart the script.